shandudata/salary_predict_model/requirements.txt
Django==1.7.4
gunicorn==19.3.0
jieba==0.35
meld3==1.0.2
mongoengine==0.10.0
numpy==1.9.2
pandas==0.16.2
Pillow==2.9.0
pymongo==3.0.3
python-dateutil==2.4.2
pytz==2015.4
scikit-learn==0.16.1
scipy==0.15.1
selenium==2.46.0
six==1.9.0
sklearn==0.0
splinter==0.7.2
supervisor==3.1.3
wheel==0.24.0
